pyproject.toml
[project]
name = "cliporaai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "aioboto3>=12.0.0",
    "asyncpg>=0.29.0",
    "bcrypt<4.0.0",
    "black>=25.1.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "celery>=5.3.6",
    "fastapi[all]>=0.116.1",
    "mypy>=1.17.0",
    "passlib>=1.7.4",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-timeout>=2.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "redis>=5.0.1",
    "ruff>=0.12.4",
    "slowapi>=0.1.9",
    "sqlalchemy-stubs>=0.4",
    "tenacity>=8.2.3",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
    "types-redis>=4.6.0.20240311",
    "boto3-stubs[s3]>=1.34.0",
    "types-aiobotocore[s3]>=2.7.0",
    "celery-types>=0.23.0",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.5",
    "pytest-timeout>=2.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "-v"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "sqlalchemy.alembic"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = "alembic.env"
ignore_missing_imports = true
ignore_errors = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "alembic.versions.*"
ignore_missing_imports = true
ignore_errors = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "alembic/versions/*"
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = "alembic/*"
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = "fastapi.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jose.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "passlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "aioboto3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "botocore.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "slowapi.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "celery.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "magic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "httpx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "starlette.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 140
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "C4", "RET", "SIM", "ARG"]
ignore = ["B008", "UP046", "B904"]  # B008: FastAPI DI; UP046: generic subclass; B904: exception chaining noise

[tool.ruff.lint.per-file-ignores]
"app/repositories/base.py" = ["UP046"]
"tests/**/*" = ["E501", "S101", "T201", "PT013", "ARG001", "ARG002", "N803", "ANN401", "S105"]
"app/core/config.py" = ["SIM105"]
"app/core/exceptions.py" = ["N818"]
"app/core/json_logging.py" = ["C404"]
"app/db/session.py" = ["ARG001"]
"app/services/storage_service.py" = ["ARG002", "RET504"]
"app/tasks/__init__.py" = ["UP035", "F401"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
suppress-none-returning = true
